refactor(lib/agents): modularize design-sub-agent.js into design-sub-agent/ modules
SD-LEO-REFACTOR-DESIGN-AGENT-001 - Refactors 1441 LOC file into 8 focused modules:
- constants.js (85 LOC) - WCAG criteria, breakpoints, defaults
- file-helpers.js (179 LOC) - File scanning utilities
- accessibility-checks.js (232 LOC) - WCAG accessibility validation
- responsive-checks.js (153 LOC) - Responsive design validation
- component-checks.js (136 LOC) - Component consistency checks
- style-checks.js (309 LOC) - Color, typography, animation checks
- design-system.js (96 LOC) - Design system compliance
- visual-verification.js (170 LOC) - Playwright MCP visual verification
- index.js (305 LOC) - DesignSubAgent class orchestration + CLI

Original file now re-exports from modular structure for backward compatibility (92 LOC).
All modules pass syntax validation. Supports independent testing of validation domains.
